action

+++ src/Bean/Bullet.java
@@ -18,14 +18,14 @@ import Util.Global;
 
 public class Bullet implements ICore{
 
-	private int positionX;
-	private int positionY;
+	private double positionX;
+	private double positionY;
 	private double speed;
 	private double angle;
 	private boolean isLive;
 	
 	
-	public Bullet(int x, int y, double speed, double angle){
+	public Bullet(double x, double y, double speed, double angle){
 		this.positionX = x;
 		this.positionY = y;
 		this.speed = speed;
@@ -64,22 +64,22 @@ public class Bullet implements ICore{
 	}
 
 	@Override
-	public void setX(int x){
+	public void setX(double x){
 		this.positionX = x;
 	}
 
 	@Override
-	public int getX(){
+	public double getX(){
 		return positionX;
 	}
 
 	@Override
-	public void setY(int y){
+	public void setY(double y){
 		this.positionY = y;
 	}
 
 	@Override
-	public int getY(){
+	public double getY(){
 		return positionY;
 	}
