+++ src/Bean/Tank.java
@@ -18,8 +18,8 @@ import Util.Global;
 
 public class Tank implements ICore{
 
-	private int positionX;
-	private int positionY;
+	private double positionX;
+	private double positionY;
 	private double speedX;
 	private double speedY;
 	private double angle;
@@ -30,7 +30,7 @@ public class Tank implements ICore{
 	private boolean direction_west;
 	private boolean direction_east;
 	
-	public Tank(int x, int y, double sx, double sy, double angle){
+	public Tank(double x, double y, double sx, double sy, double angle){
 		this.positionX = x;
 		this.positionY = y;
 		this.speedX = sx;
@@ -85,22 +85,22 @@ public class Tank implements ICore{
 	}
 	
 	@Override
-	public void setX(int x){
+	public void setX(double x){
 		this.positionX = x;
 	}
 
 	@Override
-	public int getX(){
+	public double getX(){
 		return positionX;
 	}
 
 	@Override
-	public void setY(int y){
+	public void setY(double y){
 		this.positionY = y;
 	}
 
 	@Override
-	public int getY(){
+	public double getY(){
 		return positionY;
 	}
